refactor(gui): consolidate shared UI modules
