+++ crates/openlogi-gui/src/app/home.rs
@@ -24,10 +24,10 @@ use super::AppView;
 use super::status::{loading_body, notice_body};
 use super::widgets::{add_device_button, kind_label, settings_button};
 use crate::asset::GlowGeometry;
-use crate::components::carousel::Carousel;
 use crate::features::lighting::visual as light_visual;
 use crate::state::{AppState, DeviceRecord};
-use crate::theme::{self, HEADER_H, Palette, SelectableStyle as _, Typography as _};
+use crate::ui::carousel::Carousel;
+use crate::ui::theme::{self, HEADER_H, Palette, SelectableStyle as _, Typography as _};
 
 /// Home (gallery) top bar: the "Devices" title, a Settings gear, and the
 /// Add-Device button — the entry points the old carousel header used to carry.
