--- crates/openlogi-gui/src/components/mod.rs
@@ -1,5 +1,0 @@
-//! Reusable settings-app UI components.
-
-pub mod carousel;
-pub mod device_read;
-pub mod status;
