+++ crates/openlogi-gui/src/features/action_ring.rs
@@ -17,7 +17,7 @@ use openlogi_core::binding::{ActionRingEntry, ActionRingLayout, ActionRingSlot};
 use self::action_icons::action_icon_path;
 use self::editor::action_library;
 use crate::state::AppState;
-use crate::theme::{self, Palette, Typography as _};
+use crate::ui::theme::{self, Palette, Typography as _};
 use openlogi_gui::action_ring::geometry::slot_offset;
 use openlogi_gui::action_ring::icons::ring_icon_path;
