+++ crates/openlogi-gui/src/ui/mod.rs
@@ -1,0 +1,6 @@
+//! Shared settings-app UI components and theme.
+
+pub mod carousel;
+pub mod device_read;
+pub mod status;
+pub mod theme;
